Finish splitting constants into respective files
