--- pyboy/plugins/game_wrapper_pokemon_gen1/constants.py
@@ -1,405 +1,0 @@
-from enum import Enum
-
-STRING_TERMINATOR = 80
-ASCII_DELTA = 63
-
-class Moves(Enum):
-    POUND = 0x01
-    KARATE_CHOP = 0x02
-    DOUBLESLAP = 0x03
-    COMET_PUNCH = 0x04
-    MEGA_PUNCH = 0x05
-    PAY_DAY = 0x06
-    FIRE_PUNCH = 0x07
-    ICE_PUNCH = 0x08
-    THUNDERPUNCH = 0x09
-    SCRATCH = 0x0a
-    VICEGRIP = 0x0b
-    GUILLOTINE = 0x0c
-    RAZOR_WIND = 0x0d
-    SWORDS_DANCE = 0x0e
-    CUT = 0x0f
-    GUST = 0x10
-    WING_ATTACK = 0x11
-    WHIRLWIND = 0x12
-    FLY = 0x13
-    BIND = 0x14
-    SLAM = 0x15
-    VINE_WHIP = 0x16
-    STOMP = 0x17
-    DOUBLE_KICK = 0x18
-    MEGA_KICK = 0x19
-    JUMP_KICK = 0x1a
-    ROLLING_KICK = 0x1b
-    SAND_ATTACK = 0x1c
-    HEADBUTT = 0x1d
-    HORN_ATTACK = 0x1e
-    FURY_ATTACK = 0x1f
-    HORN_DRILL = 0x20
-    TACKLE = 0x21
-    BODY_SLAM = 0x22
-    WRAP = 0x23
-    TAKE_DOWN = 0x24
-    THRASH = 0x25
-    DOUBLE_EDGE = 0x26
-    TAIL_WHIP = 0x27
-    POISON_STING = 0x28
-    TWINEEDLE = 0x29
-    PIN_MISSILE = 0x2a
-    LEER = 0x2b
-    BITE = 0x2c
-    GROWL = 0x2d
-    ROAR = 0x2e
-    SING = 0x2f
-    SUPERSONIC = 0x30
-    SONICBOOM = 0x31
-    DISABLE = 0x32
-    ACID = 0x33
-    EMBER = 0x34
-    FLAMETHROWER = 0x35
-    MIST = 0x36
-    WATER_GUN = 0x37
-    HYDRO_PUMP = 0x38
-    SURF = 0x39
-    ICE_BEAM = 0x3a
-    BLIZZARD = 0x3b
-    PSYBEAM = 0x3c
-    BUBBLEBEAM = 0x3d
-    AURORA_BEAM = 0x3e
-    HYPER_BEAM = 0x3f
-    PECK = 0x40
-    DRILL_PECK = 0x41
-    SUBMISSION = 0x42
-    LOW_KICK = 0x43
-    COUNTER = 0x44
-    SEISMIC_TOSS = 0x45
-    STRENGTH = 0x46
-    ABSORB = 0x47
-    MEGA_DRAIN = 0x48
-    LEECH_SEED = 0x49
-    GROWTH = 0x4a
-    RAZOR_LEAF = 0x4b
-    SOLARBEAM = 0x4c
-    POISONPOWDER = 0x4d
-    STUN_SPORE = 0x4e
-    SLEEP_POWDER = 0x4f
-    PETAL_DANCE = 0x50
-    STRING_SHOT = 0x51
-    DRAGON_RAGE = 0x52
-    FIRE_SPIN = 0x53
-    THUNDERSHOCK = 0x54
-    THUNDERBOLT = 0x55
-    THUNDER_WAVE = 0x56
-    THUNDER = 0x57
-    ROCK_THROW = 0x58
-    EARTHQUAKE = 0x59
-    FISSURE = 0x5a
-    DIG = 0x5b
-    TOXIC = 0x5c
-    CONFUSION = 0x5d
-    PSYCHIC_M = 0x5e
-    HYPNOSIS = 0x5f
-    MEDITATE = 0x60
-    AGILITY = 0x61
-    QUICK_ATTACK = 0x62
-    RAGE = 0x63
-    TELEPORT = 0x64
-    NIGHT_SHADE = 0x65
-    MIMIC = 0x66
-    SCREECH = 0x67
-    DOUBLE_TEAM = 0x68
-    RECOVER = 0x69
-    HARDEN = 0x6a
-    MINIMIZE = 0x6b
-    SMOKESCREEN = 0x6c
-    CONFUSE_RAY = 0x6d
-    WITHDRAW = 0x6e
-    DEFENSE_CURL = 0x6f
-    BARRIER = 0x70
-    LIGHT_SCREEN = 0x71
-    HAZE = 0x72
-    REFLECT = 0x73
-    FOCUS_ENERGY = 0x74
-    BIDE = 0x75
-    METRONOME = 0x76
-    MIRROR_MOVE = 0x77
-    SELFDESTRUCT = 0x78
-    EGG_BOMB = 0x79
-    LICK = 0x7a
-    SMOG = 0x7b
-    SLUDGE = 0x7c
-    BONE_CLUB = 0x7d
-    FIRE_BLAST = 0x7e
-    WATERFALL = 0x7f
-    CLAMP = 0x80
-    SWIFT = 0x81
-    SKULL_BASH = 0x82
-    SPIKE_CANNON = 0x83
-    CONSTRICT = 0x84
-    AMNESIA = 0x85
-    KINESIS = 0x86
-    SOFTBOILED = 0x87
-    HI_JUMP_KICK = 0x88
-    GLARE = 0x89
-    DREAM_EATER = 0x8a
-    POISON_GAS = 0x8b
-    BARRAGE = 0x8c
-    LEECH_LIFE = 0x8d
-    LOVELY_KISS = 0x8e
-    SKY_ATTACK = 0x8f
-    TRANSFORM = 0x90
-    BUBBLE = 0x91
-    DIZZY_PUNCH = 0x92
-    SPORE = 0x93
-    FLASH = 0x94
-    PSYWAVE = 0x95
-    SPLASH = 0x96
-    ACID_ARMOR = 0x97
-    CRABHAMMER = 0x98
-    EXPLOSION = 0x99
-    FURY_SWIPES = 0x9a
-    BONEMERANG = 0x9b
-    REST = 0x9c
-    ROCK_SLIDE = 0x9d
-    HYPER_FANG = 0x9e
-    SHARPEN = 0x9f
-    CONVERSION = 0xa0
-    TRI_ATTACK = 0xa1
-    SUPER_FANG = 0xa2
-    SLASH = 0xa3
-    SUBSTITUTE = 0xa4
-
-class Items(Enum):
-    MASTER_BALL = 0x01
-    ULTRA_BALL = 0x02
-    GREAT_BALL = 0x03
-    POKE_BALL = 0x04
-    TOWN_MAP = 0x05
-    BICYCLE = 0x06
-    SURFBOARD = 0x07
-    SAFARI_BALL = 0x08
-    POKEDEX = 0x09
-    MOON_STONE = 0x0A
-    ANTIDOTE = 0x0B
-    BURN_HEAL = 0x0C
-    ICE_HEAL = 0x0D
-    AWAKENING = 0x0E
-    PARLYZ_HEAL = 0x0F
-    FULL_RESTORE = 0x10
-    MAX_POTION = 0x11
-    HYPER_POTION = 0x12
-    SUPER_POTION = 0x13
-    POTION = 0x14
-    BOULDERBADGE = 0x15
-    CASCADEBADGE = 0x16
-    THUNDERBADGE = 0x17
-    RAINBOWBADGE = 0x18
-    SOULBADGE = 0x19
-    MARSHBADGE = 0x1A
-    VOLCANOBADGE = 0x1B
-    EARTHBADGE = 0x1C
-    ESCAPE_ROPE = 0x1D
-    REPEL = 0x1E
-    OLD_AMBER = 0x1F
-    FIRE_STONE = 0x20
-    THUNDER_STONE = 0x21
-    WATER_STONE = 0x22
-    HP_UP = 0x23
-    PROTEIN = 0x24
-    IRON = 0x25
-    CARBOS = 0x26
-    CALCIUM = 0x27
-    RARE_CANDY = 0x28
-    DOME_FOSSIL = 0x29
-    HELIX_FOSSIL = 0x2A
-    SECRET_KEY = 0x2B
-    UNUSED_ITEM = 0x2C
-    BIKE_VOUCHER = 0x2D
-    X_ACCURACY = 0x2E
-    LEAF_STONE = 0x2F
-    CARD_KEY = 0x30
-    NUGGET = 0x31
-    PP_UP_2 = 0x32
-    POKE_DOLL = 0x33
-    FULL_HEAL = 0x34
-    REVIVE = 0x35
-    MAX_REVIVE = 0x36
-    GUARD_SPEC = 0x37
-    SUPER_REPEL = 0x38
-    MAX_REPEL = 0x39
-    DIRE_HIT = 0x3A
-    COIN = 0x3B
-    FRESH_WATER = 0x3C
-    SODA_POP = 0x3D
-    LEMONADE = 0x3E
-    S_S_TICKET = 0x3F
-    GOLD_TEETH = 0x40
-    X_ATTACK = 0x41
-    X_DEFEND = 0x42
-    X_SPEED = 0x43
-    X_SPECIAL = 0x44
-    COIN_CASE = 0x45
-    OAKS_PARCEL = 0x46
-    ITEMFINDER = 0x47
-    SILPH_SCOPE = 0x48
-    POKE_FLUTE = 0x49
-    LIFT_KEY = 0x4A
-    EXP_ALL = 0x4B
-    OLD_ROD = 0x4C
-    GOOD_ROD = 0x4D
-    SUPER_ROD = 0x4E
-    PP_UP = 0x4F
-    ETHER = 0x50
-    MAX_ETHER = 0x51
-    ELIXER = 0x52
-    MAX_ELIXER = 0x53
-    FLOOR_B2F = 0x54
-    FLOOR_B1F = 0x55
-    FLOOR_1F = 0x56
-    FLOOR_2F = 0x57
-    FLOOR_3F = 0x58
-    FLOOR_4F = 0x59
-    FLOOR_5F = 0x5A
-    FLOOR_6F = 0x5B
-    FLOOR_7F = 0x5C
-    FLOOR_8F = 0x5D
-    FLOOR_9F = 0x5E
-    FLOOR_10F = 0x5F
-    FLOOR_11F = 0x60
-    FLOOR_B4F = 0x61
-    HM_01 = 0xC4
-    HM_02 = 0xC5
-    HM_03 = 0xC6
-    HM_04 = 0xC7
-    HM_05 = 0xC8
-    TM_01 = 0xC9
-    TM_02 = 0xCA
-    TM_03 = 0xCB
-    TM_04 = 0xCC
-    TM_05 = 0xCD
-    TM_06 = 0xCE
-    TM_07 = 0xCF
-    TM_08 = 0xD0
-    TM_09 = 0xD1
-    TM_10 = 0xD2
-    TM_11 = 0xD3
-    TM_12 = 0xD4
-    TM_13 = 0xD5
-    TM_14 = 0xD6
-    TM_15 = 0xD7
-    TM_16 = 0xD8
-    TM_17 = 0xD9
-    TM_18 = 0xDA
-    TM_19 = 0xDB
-    TM_20 = 0xDC
-    TM_21 = 0xDD
-    TM_22 = 0xDE
-    TM_23 = 0xDF
-    TM_24 = 0xE0
-    TM_25 = 0xE1
-    TM_26 = 0xE2
-    TM_27 = 0xE3
-    TM_28 = 0xE4
-    TM_29 = 0xE5
-    TM_30 = 0xE6
-    TM_31 = 0xE7
-    TM_32 = 0xE8
-    TM_33 = 0xE9
-    TM_34 = 0xEA
-    TM_35 = 0xEB
-    TM_36 = 0xEC
-    TM_37 = 0xED
-    TM_38 = 0xEE
-    TM_39 = 0xEF
-    TM_40 = 0xF0
-    TM_41 = 0xF1
-    TM_42 = 0xF2
-    TM_43 = 0xF3
-    TM_44 = 0xF4
-    TM_45 = 0xF5
-    TM_46 = 0xF6
-    TM_47 = 0xF7
-    TM_48 = 0xF8
-    TM_49 = 0xF9
-    TM_50 = 0xFA
-
-class Sprites(Enum):
-    RED = 0x01
-    BLUE = 0x02
-    OAK = 0x03
-    BUG_CATCHER = 0x04
-    SLOWBRO = 0x05
-    LASS = 0x06
-    BLACK_HAIR_BOY_1 = 0x07
-    LITTLE_GIRL = 0x08
-    BIRD = 0x09
-    FAT_BALD_GUY = 0x0a
-    GAMBLER = 0x0b
-    BLACK_HAIR_BOY_2 = 0x0c
-    GIRL = 0x0d
-    HIKER = 0x0e
-    FOULARD_WOMAN = 0x0f
-    GENTLEMAN = 0x10
-    DAISY = 0x11
-    BIKER = 0x12
-    SAILOR = 0x13
-    COOK = 0x14
-    BIKE_SHOP_GUY = 0x15
-    MR_FUJI = 0x16
-    GIOVANNI = 0x17
-    ROCKET = 0x18
-    MEDIUM = 0x19
-    WAITER = 0x1a
-    ERIKA = 0x1b
-    MOM_GEISHA = 0x1c
-    BRUNETTE_GIRL = 0x1d
-    LANCE = 0x1e
-    OAK_SCIENTIST_AIDE = 0x1f
-    OAK_AIDE = 0x20
-    ROCKER = 0x21
-    SWIMMER = 0x22
-    WHITE_PLAYER = 0x23
-    GYM_HELPER = 0x24
-    OLD_PERSON = 0x25
-    MART_GUY = 0x26
-    FISHER = 0x27
-    OLD_MEDIUM_WOMAN = 0x28
-    NURSE = 0x29
-    CABLE_CLUB_WOMAN = 0x2a
-    MR_MASTERBALL = 0x2b
-    LAPRAS_GIVER = 0x2c
-    WARDEN = 0x2d
-    SS_CAPTAIN = 0x2e
-    FISHER2 = 0x2f
-    BLACKBELT = 0x30
-    GUARD = 0x31
-    COP_GUARD = 0x32
-    MOM = 0x33
-    BALDING_GUY = 0x34
-    YOUNG_BOY = 0x35
-    GAMEBOY_KID = 0x36
-    GAMEBOY_KID_COPY = 0x37
-    CLEFAIRY = 0x38
-    AGATHA = 0x39
-    BRUNO = 0x3a
-    LORELEI = 0x3b
-    SEEL = 0x3c
-    BALL = 0x3d
-    OMANYTE = 0x3e
-    BOULDER = 0x3f
-    PAPER_SHEET = 0x40
-    BOOK_MAP_DEX = 0x41
-    CLIPBOARD = 0x42
-    SNORLAX = 0x43
-    OLD_AMBER_COPY = 0x44
-    OLD_AMBER = 0x45
-    LYING_OLD_MAN_UNUSED_1 = 0x46
-    LYING_OLD_MAN_UNUSED_2 = 0x47
-    LYING_OLD_MAN = 0x48
-
-class Statuses(Enum):
-    ASLEEP = 4
-    POISONED = 8
-    BURNED = 16
-    FROZEN = 32
-    PARALYZED = 64
