+++ pyboy/plugins/game_wrapper_pokemon_gen1/data/constants/sprites.py
@@ -1,0 +1,75 @@
+from enum import Enum
+
+class Sprites(Enum):
+    RED = 0x01
+    BLUE = 0x02
+    OAK = 0x03
+    BUG_CATCHER = 0x04
+    SLOWBRO = 0x05
+    LASS = 0x06
+    BLACK_HAIR_BOY_1 = 0x07
+    LITTLE_GIRL = 0x08
+    BIRD = 0x09
+    FAT_BALD_GUY = 0x0a
+    GAMBLER = 0x0b
+    BLACK_HAIR_BOY_2 = 0x0c
+    GIRL = 0x0d
+    HIKER = 0x0e
+    FOULARD_WOMAN = 0x0f
+    GENTLEMAN = 0x10
+    DAISY = 0x11
+    BIKER = 0x12
+    SAILOR = 0x13
+    COOK = 0x14
+    BIKE_SHOP_GUY = 0x15
+    MR_FUJI = 0x16
+    GIOVANNI = 0x17
+    ROCKET = 0x18
+    MEDIUM = 0x19
+    WAITER = 0x1a
+    ERIKA = 0x1b
+    MOM_GEISHA = 0x1c
+    BRUNETTE_GIRL = 0x1d
+    LANCE = 0x1e
+    OAK_SCIENTIST_AIDE = 0x1f
+    OAK_AIDE = 0x20
+    ROCKER = 0x21
+    SWIMMER = 0x22
+    WHITE_PLAYER = 0x23
+    GYM_HELPER = 0x24
+    OLD_PERSON = 0x25
+    MART_GUY = 0x26
+    FISHER = 0x27
+    OLD_MEDIUM_WOMAN = 0x28
+    NURSE = 0x29
+    CABLE_CLUB_WOMAN = 0x2a
+    MR_MASTERBALL = 0x2b
+    LAPRAS_GIVER = 0x2c
+    WARDEN = 0x2d
+    SS_CAPTAIN = 0x2e
+    FISHER2 = 0x2f
+    BLACKBELT = 0x30
+    GUARD = 0x31
+    COP_GUARD = 0x32
+    MOM = 0x33
+    BALDING_GUY = 0x34
+    YOUNG_BOY = 0x35
+    GAMEBOY_KID = 0x36
+    GAMEBOY_KID_COPY = 0x37
+    CLEFAIRY = 0x38
+    AGATHA = 0x39
+    BRUNO = 0x3a
+    LORELEI = 0x3b
+    SEEL = 0x3c
+    BALL = 0x3d
+    OMANYTE = 0x3e
+    BOULDER = 0x3f
+    PAPER_SHEET = 0x40
+    BOOK_MAP_DEX = 0x41
+    CLIPBOARD = 0x42
+    SNORLAX = 0x43
+    OLD_AMBER_COPY = 0x44
+    OLD_AMBER = 0x45
+    LYING_OLD_MAN_UNUSED_1 = 0x46
+    LYING_OLD_MAN_UNUSED_2 = 0x47
+    LYING_OLD_MAN = 0x48
