+++ pyboy/plugins/game_wrapper_pokemon_gen1/data/constants/status.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+class Statuses(Enum):
+    ASLEEP = 4
+    POISONED = 8
+    BURNED = 16
+    FROZEN = 32
+    PARALYZED = 64
